requirements.txt
# Core dependencies for Claude Agent SDK in E2B sandboxes
claude-agent-sdk>=0.1.0
e2b>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Optional dependencies for examples
fastapi>=0.110.0  # For API server example
uvicorn>=0.27.0   # For API server example
pydantic>=2.0.0   # For API server example

# Development dependencies
# Add these if you're contributing:
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# mypy>=1.8.0
# ruff>=0.2.0
